+++ src/gen_rpc_driver.erl
@@ -14,7 +14,7 @@
 
 -callback activate_socket(term()) -> ok.
 
--callback authenticate_server(term()) -> ok | {error, {badtcp | badrpc, term()}}.
+-callback authenticate_to_server(atom(), term()) -> ok | {error, {badtcp | badrpc, term()}}.
 
 -callback authenticate_client(term(), tuple(), binary()) -> ok | {error, {badtcp | badrpc, term()}}.