+++ src/supervisor/gen_rpc_sup.erl
@@ -30,7 +30,8 @@ init([]) ->
     {ok, {{one_for_one, 100, 1}, [
         {gen_rpc_server_tcp, {gen_rpc_server,start_link,[tcp]}, permanent, 5000, worker, [gen_rpc_server]},
         {gen_rpc_server_ssl, {gen_rpc_server,start_link,[ssl]}, permanent, 5000, worker, [gen_rpc_server]},
-        {gen_rpc_acceptor_sup, {gen_rpc_acceptor_sup,start_link, []}, permanent, 5000, supervisor, [gen_rpc_acceptor_sup]},
-        {gen_rpc_dispatcher, {gen_rpc_dispatcher,start_link, []}, permanent, 5000, worker, [gen_rpc_dispatcher]},
-        {gen_rpc_client_sup, {gen_rpc_client_sup,start_link, []}, permanent, 5000, supervisor, [gen_rpc_client_sup]}
+        {gen_rpc_acceptor_sup, {gen_rpc_acceptor_sup,start_link,[]}, permanent, 5000, supervisor, [gen_rpc_acceptor_sup]},
+        {gen_rpc_monitor, {gen_rpc_monitor,start_link,[]}, permanent, 5000, worker, [gen_rpc_monitor]},
+        {gen_rpc_dispatcher, {gen_rpc_dispatcher,start_link,[]}, permanent, 5000, worker, [gen_rpc_dispatcher]},
+        {gen_rpc_client_sup, {gen_rpc_client_sup,start_link,[]}, permanent, 5000, supervisor, [gen_rpc_client_sup]}
     ]}}.