+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Below is a non-exhaustive list of changes between `gen_rpc` versions.
 - Deprecate support for Erlang < 21.0
 - Support monitoring nodes
 - Support EC SSL certificates
+- Support cookie per node configuration
+- Support external cookie validation mechanism
+- Support keepalive gen_server that actively keeps a client connection alive
 
 ## 2.1.0